ad74416h: migrate to CMake/Kconfig and drop Mbed
Convert the ad74416h examples from the legacy Make stack to the new
CMake/Kconfig stack. Drop the legacy Make file, various .mk files and
builds.json file.

Remove platform support for Mbed due to EOL.
