--- projects/ad74416h/Makefile
@@ -1,9 +1,0 @@
-EXAMPLE ?= basic
-
-include ../../tools/scripts/generic_variables.mk
-
-include ../../tools/scripts/examples.mk
-
-include src.mk
-
-include ../../tools/scripts/generic.mk
