+++ projects/ad74416h/boards/current_input_loop/sdp-ck1z.conf
@@ -1,0 +1,1 @@
+CONFIG_STM32_IOC_PATH="sdp-ck1z.ioc"
