+++ projects/ad74416h/boards/multiple_devices/sdp-ck1z.conf
@@ -1,0 +1,1 @@
+CONFIG_STM32_IOC_PATH="sdp-ck1z.ioc"
