+++ projects/ad74416h/boards/temperature_3wire_rtd/sdp-ck1z.conf
@@ -1,0 +1,1 @@
+CONFIG_STM32_IOC_PATH="sdp-ck1z.ioc"
