--- projects/ad74416h/builds.json
@@ -1,43 +1,0 @@
-{
-  "mbed": {
-    "basic_example": {
-      "flags" : "EXAMPLE=basic"
-    },
-    "current_input_ext": {
-      "flags" : "EXAMPLE=current_input_ext"
-    },
-    "current_input_loop": {
-      "flags" : "EXAMPLE=current_input_loop"
-    },
-    "current_output": {
-      "flags" : "EXAMPLE=current_output"
-    },
-    "multiple_devices": {
-      "flags" : "EXAMPLE=multiple_devices"
-    },
-    "digital_input_logic": {
-      "flags" : "EXAMPLE=digital_input_logic"
-    },
-    "digital_input_loop": {
-      "flags" : "EXAMPLE=digital_input_loop"
-    },
-    "digital_output": {
-      "flags" : "EXAMPLE=digital_output"
-    },
-    "temperature_2wire_rtd": {
-      "flags" : "EXAMPLE=temperature_2wire_rtd"
-    },
-    "temperature_3wire_rtd": {
-      "flags" : "EXAMPLE=temperature_3wire_rtd"
-    },  
-    "voltage_input": {
-      "flags" : "EXAMPLE=voltage_input"
-    },
-    "voltage_input_irq": {
-      "flags" : "EXAMPLE=voltage_input_irq"
-    },
-    "voltage_output": {
-      "flags" : "EXAMPLE=voltage_output"
-    }
-  }
-}
