+++ projects/ad74416h/temperature_2wire_rtd.conf
@@ -1,0 +1,7 @@
+CONFIG_UART=y
+CONFIG_SPI=y
+CONFIG_GPIO=y
+CONFIG_IRQ=y
+CONFIG_ADC_DAC=y
+CONFIG_ADC_DAC_AD74416H=y
+CONFIG_AD74416H_TEMPERATURE_2WIRE_RTD_EXAMPLE=y
