+++ projects/ad74416h/voltage_input_irq.conf
@@ -1,0 +1,7 @@
+CONFIG_UART=y
+CONFIG_SPI=y
+CONFIG_GPIO=y
+CONFIG_IRQ=y
+CONFIG_ADC_DAC=y
+CONFIG_ADC_DAC_AD74416H=y
+CONFIG_AD74416H_VOLTAGE_INPUT_IRQ_EXAMPLE=y
